refonte profil 2

+++ app/Plugin/AuthAcl/Controller/UsersController.php
@@ -115,8 +115,10 @@ class UsersController extends AuthAclAppController {
 
 
 	public function editAccount(){
-		$this->set('title_for_layout', __('Mon compte'));
+
 		$this->layout = 'default';
+		$this->set('title_for_layout', __('Mon compte'));
+
 		if ($this->request->isAjax()){
 			$this->autoRender = false;
 			$this->layout = null;
@@ -133,7 +135,10 @@ class UsersController extends AuthAclAppController {
 			throw new NotFoundException(__('Invalid user'));
 		}
 		if ($this->request->is('post') || $this->request->is('put')) {
-		 	unlink(WWW_ROOT . DS .  files . DS . users. DS . thumbnails .DS . $id .'_scale_150' . '.jpg' );
+			if(file_exists(WWW_ROOT . DS .  files . DS . users. DS . thumbnails .DS . $id .'_scale_150' . '.jpg'))
+			{
+		 		unlink(WWW_ROOT . DS .  files . DS . users. DS . thumbnails .DS . $id .'_scale_150' . '.jpg' );
+			}
 			if (empty($this->request->data['User']['user_password'])){
 				unset($this->request->data['User']['user_password']);
 				unset($this->User->validate['user_confirm_password']['checkPassword']);
@@ -151,6 +156,8 @@ class UsersController extends AuthAclAppController {
 				$errors = $this->User->validationErrors;
 			}
 		} else {
+
+			$this->layout = 'default';
 			$this->request->data = am($this->request->data,$this->User->read(null, $id));
 			unset($this->request->data['User']['user_password']);
 		}
@@ -707,9 +714,6 @@ class UsersController extends AuthAclAppController {
 				}
 			}
 		}
-
-
-
 		$this->redirect(array('action' => 'activecomplete'));
 
 	}
